web-app-rsa/cart/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
httpx==0.28.1
PyJWT==2.10.1
pydantic==2.10.6
cryptography==44.0.2
